Add metrics to measture the time a task waiting in history queue (#6205)
